Fix create migration for sidekiqs table

+++ resources/migrations/20151021211459-create-sidekiqs-table.up.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXISTS sidekiqs(id BIGINT PRIMARY KEY NOT NULL, redis_instance TEXT, namespace TEXT, application TEXT);
+CREATE TABLE IF NOT EXISTS sidekiqs(id SERIAL PRIMARY KEY NOT NULL, redis_instance TEXT, namespace TEXT, application TEXT);
